en migration modulos he cambiado horas de tinyInteger a integer ya que no podia almacenar mas de 127 horas

+++ database/migrations/2024_12_09_120022_create_modulos_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         $table->id();
         $table->string('codigo', 10)->unique();
         $table->string('nombre', 100);
-        $table->tinyInteger('horas');
+        $table->integer('horas');
         $table->tinyInteger('curso');
         $table->unsignedBigInteger('id_ciclo');
         $table->timestamps();
